Ignore sorts on bogus columns
